fix diagramm and legends values = now equal
