+++ app/internal/methodsForAnalytic/methodsForExpenses/expDiagrams.go
@@ -28,12 +28,12 @@ func GenerateWeeklyExpensePieChartURL(categorySummary map[string]uint64) (string
 		"Прочее":              "#D3D3D3", // серый
 	}
 
-	// Общий расход
+	// считаем общий расход
 	for _, value := range categorySummary {
 		totalExpense += value
 	}
 
-	// Преобразуем суммы в проценты
+	// преобразуем суммы в проценты
 	for category, value := range categorySummary {
 		if value > 0 {
 			labels = append(labels, category)
@@ -42,7 +42,7 @@ func GenerateWeeklyExpensePieChartURL(categorySummary map[string]uint64) (string
 		}
 	}
 
-	// Присваиваем цвета
+	// присваиваем цвета
 	for _, category := range labels {
 		if color, exists := categoryColors[category]; exists {
 			colors = append(colors, color)
@@ -50,7 +50,7 @@ func GenerateWeeklyExpensePieChartURL(categorySummary map[string]uint64) (string
 			colors = append(colors, "#CCCCCC") // цвет по умолчанию
 		}
 	}
-
+	// заполнение для квикчарта
 	chartData := map[string]interface{}{
 		"type": "doughnut", // Тип графика
 		"data": map[string]interface{}{
@@ -71,13 +71,13 @@ func GenerateWeeklyExpensePieChartURL(categorySummary map[string]uint64) (string
 		},
 	}
 
-	// Преобразуем в JSON
+	// преобразуем в JSON
 	jsonData, err := json.Marshal(chartData)
 	if err != nil {
 		return "", fmt.Errorf("Ошибка при создании данных для диаграммы: %v", err)
 	}
 
-	// Составляем URL
+	// составляем URL
 	baseURL := "https://quickchart.io/chart"
 	params := url.Values{}
 	params.Add("c", string(jsonData))
