+++ app/pkg/TgBot/pushOnAnalyticButtons.go
@@ -1,11 +1,9 @@
 package TgBot
 
 import (
-	"cachManagerApp/app/db/models"
 	"cachManagerApp/app/internal/methodsForAnalytic/methodsForSummary"
 	"cachManagerApp/app/internal/notion"
 	"cachManagerApp/app/pkg/ButtonsCreate"
-	"cachManagerApp/database"
 	"log/slog"
 
 	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
@@ -64,14 +62,4 @@ func PushOnAnalyticButton(bot *tgbotapi.BotAPI, update tgbotapi.Update, buttonCr
 			log.Error("Error sending message: %v", err)
 		}
 	}
-}
-
-// Получение валюты из бд
-func CurrencyFromChatID(chatID int64) (string, error) {
-	var user models.Users
-	result := database.DB.Where("telegram_id = ?", chatID).First(&user)
-	if result.Error != nil {
-		return "", result.Error
-	}
-	return user.Currency, nil
 }
